feat: crear partida como hijo de una subcategoria
