+++ src/app/(admin)/execution-project/budgets/work-item/CreateWorkItemDialog.tsx
@@ -1,7 +1,6 @@
 import { useWorkItem } from "@/hooks/use-workitem";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Plus } from "lucide-react";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
@@ -12,7 +11,6 @@ import {
     DialogDescription,
     DialogHeader,
     DialogTitle,
-    DialogTrigger,
 } from "@/components/ui/dialog";
 import {
     Form,
@@ -26,17 +24,17 @@ import {
 import { Input } from "@/components/ui/input";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 
-export function CreateCategoryDialog() {
-    const [open, setOpen] = useState(false);
-
+export function CreateWorkItemDialog({
+    open,
+    onOpenChange,
+    subcategoryId,
+}: {
+    open: boolean;
+    onOpenChange: (open: boolean) => void;
+    subcategoryId: string;
+}) {
     return (
-        <Dialog open={open} onOpenChange={setOpen}>
-            <DialogTrigger asChild>
-                <Button variant="outline" size="sm">
-                    <Plus className="mr-2 size-4" aria-hidden="true" />
-                    Crear Categoría
-                </Button>
-            </DialogTrigger>
+        <Dialog open={open} onOpenChange={onOpenChange}>
             <DialogContent>
                 <DialogHeader>
                     <DialogTitle>Crear una partida</DialogTitle>
@@ -54,10 +52,16 @@ export function CreateCategoryDialog() {
                     </TabsList>
 
                     <TabsContent value="apu" className="space-y-4">
-                        <CreateWithApuForm setOpen={setOpen} />
+                        <CreateWithApuForm
+                            setOpen={onOpenChange}
+                            subcategoryId={subcategoryId}
+                        />
                     </TabsContent>
                     <TabsContent value="subworkitem">
-                        <CreateWithSubitemsForm setOpen={setOpen} />
+                        <CreateWithSubitemsForm
+                            setOpen={onOpenChange}
+                            subcategoryId={subcategoryId}
+                        />
                     </TabsContent>
                 </Tabs>
             </DialogContent>
@@ -81,7 +85,14 @@ const withApuSchema = z.object({
     apuPerformance: z.coerce.number().min(0),
     apuWorkHours: z.coerce.number().min(0),
 });
-function CreateWithApuForm({ setOpen }: { setOpen: (v: boolean) => void }) {
+
+function CreateWithApuForm({
+    setOpen,
+    subcategoryId,
+}: {
+    setOpen: (v: boolean) => void;
+    subcategoryId: string;
+}) {
     const { onCreate, createLoading, createSuccess } = useWorkItem();
     const form = useForm<z.infer<typeof withApuSchema>>({
         resolver: zodResolver(withApuSchema),
@@ -94,8 +105,10 @@ function CreateWithApuForm({ setOpen }: { setOpen: (v: boolean) => void }) {
             apuWorkHours: "",
         },
     });
+
     function onSubmit(values: z.infer<typeof withApuSchema>) {
         onCreate({
+            subcategoryId,
             name: values.name,
             unit: values.unit,
             apu: {
@@ -222,10 +235,13 @@ const withSubitemsSchema = z.object({
         })
         .max(50),
 });
+
 function CreateWithSubitemsForm({
     setOpen,
+    subcategoryId,
 }: {
     setOpen: (v: boolean) => void;
+    subcategoryId: string;
 }) {
     const { onCreate, createLoading, createSuccess } = useWorkItem();
     const form = useForm<z.infer<typeof withSubitemsSchema>>({
@@ -236,6 +252,7 @@ function CreateWithSubitemsForm({
     });
     function onSubmit(values: z.infer<typeof withSubitemsSchema>) {
         onCreate({
+            subcategoryId,
             name: values.name,
         });
     }
